Added class hierarchy to relate game modes.
