+++ CMakeLists.txt
@@ -30,6 +30,7 @@ set(SOURCES
     ./src/NextQueue.cpp
     ./src/Controller.cpp
     ./src/SinglePlayerGame.cpp
+    ./src/TwoPlayerGame.cpp
     ./src/Raytris.cpp
     ./src/main.cpp
 )
